feat: hashs name length with yaml
