+++ app/services/testyaml.rb
@@ -1,0 +1,4 @@
+require 'yaml'
+
+yaml_data = YAML.load_file('exemplo1.yaml')
+p yaml_data
